refactor: consolidate deck form into page
