+++ src/pages/deck-form/ui/DeckFormPage.tsx
@@ -1,13 +1,46 @@
 import type * as React from "react";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 
-import { DeckFormContainer } from "./DeckFormContainer";
+import { DeckDeletionDialog, useDeckDeletion } from "@/features/deck-deletion";
+import { routes } from "@/shared/router";
+import { AppLayout } from "@/widgets/app-layout";
+import { RouteNotFound } from "@/widgets/route-not-found";
+
+import { useDeckFormState } from "../model/useDeckFormState";
+import { DeckEditor } from "./DeckEditor";
+
+const DeckFormContent: React.FC<{ deckId: string }> = ({ deckId }) => {
+  const navigate = useNavigate();
+  const goToList = () => void navigate(routes.deckList.to(), { replace: true });
+  const editor = useDeckFormState({ deckId, onCancel: goToList, onSaved: goToList });
+  const deletion = useDeckDeletion({ onDeleted: goToList });
+
+  if (editor == null) {
+    return (
+      <RouteNotFound title="Deck not found" description="The requested deck is unavailable or has been removed." />
+    );
+  }
+
+  return (
+    <AppLayout showHeader>
+      {deletion.target != null && (
+        <DeckDeletionDialog target={deletion.target} onCancel={deletion.cancel} onConfirm={deletion.confirm} />
+      )}
+      <DeckEditor
+        deckName={editor.deckName}
+        form={editor.form}
+        saveError={editor.saveError}
+        onDelete={() => deletion.request(editor.form.deckInfo.id)}
+      />
+    </AppLayout>
+  );
+};
 
 export const DeckFormPage: React.FC = () => {
   const params = useParams();
   const deckId = params.id;
   if (deckId == null) throw new Error("invalid deck id");
 
-  // Form state belongs to one route Deck and must reset when the id changes.
-  return <DeckFormContainer key={deckId} deckId={deckId} />;
+  // Page-owned form and deletion state must not survive navigation to a different Deck.
+  return <DeckFormContent key={deckId} deckId={deckId} />;
 };
